requirements.txt
pytest>=7.4.0
tqdm>=4.66.0
plotly>=5.17.0
pandas>=2.1.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

